extraction/requirements.txt
beautifulsoup4==4.9.3
pyspark==3.1.2
chispa==0.8.2
pytest==6.2.4